+++ pkgs/native_assets_cli/lib/src/model/asset.dart
@@ -14,8 +14,6 @@ abstract class AssetPath {
     switch (pathType) {
       case AssetAbsolutePath._pathTypeValue:
         return AssetAbsolutePath(uri!);
-      case AssetRelativePath._pathTypeValue:
-        return AssetRelativePath(uri!);
       case AssetSystemPath._pathTypeValue:
         return AssetSystemPath(uri!);
       case AssetInExecutable._pathTypeValue:
@@ -65,41 +63,6 @@ class AssetAbsolutePath implements AssetPath {
   @override
   bool operator ==(Object other) {
     if (other is! AssetAbsolutePath) {
-      return false;
-    }
-    return uri == other.uri;
-  }
-
-  @override
-  Future<bool> exists() => uri.fileSystemEntity.exists();
-}
-
-/// Asset is avaliable on a relative path.
-///
-/// If [LinkMode] of an [Asset] is [LinkMode.dynamic],
-/// `Platform.script.resolve(uri)` will be used to load the asset at runtime.
-class AssetRelativePath implements AssetPath {
-  final Uri uri;
-
-  AssetRelativePath(this.uri);
-
-  static const _pathTypeValue = 'relative';
-
-  @override
-  Map<String, Object> toYaml() => {
-        AssetPath._pathTypeKey: _pathTypeValue,
-        AssetPath._uriKey: uri.toFilePath(),
-      };
-
-  @override
-  List<String> toDartConst() => [_pathTypeValue, uri.toFilePath()];
-
-  @override
-  int get hashCode => Object.hash(uri, 133717);
-
-  @override
-  bool operator ==(Object other) {
-    if (other is! AssetRelativePath) {
       return false;
     }
     return uri == other.uri;
@@ -263,8 +226,6 @@ class Asset {
         id: path.toDartConst(),
       };
 
-  String toYamlString() => yamlEncode(toYaml());
-
   static const _idKey = 'id';
   static const _linkModeKey = 'link_mode';
   static const _pathKey = 'path';
@@ -280,44 +241,4 @@ extension AssetIterable on Iterable<Asset> {
   List<Object> toYaml() => [for (final item in this) item.toYaml()];
 
   String toYamlString() => yamlEncode(toYaml());
-
-  Iterable<Asset> whereLinkMode(LinkMode linkMode) =>
-      where((e) => e.linkMode == linkMode);
-
-  Map<Target, List<Asset>> get assetsPerTarget {
-    final result = <Target, List<Asset>>{};
-    for (final asset in this) {
-      final assets = result[asset.target] ?? [];
-      assets.add(asset);
-      result[asset.target] = assets;
-    }
-    return result;
-  }
-
-  Map<String, Map<String, List<String>>> toDartConst() => {
-        for (final entry in assetsPerTarget.entries)
-          entry.key.toString():
-              _combineMaps(entry.value.map((e) => e.toDartConst()).toList())
-      };
-
-  Map<Object, Object> toNativeAssetsFileEncoding() => {
-        'format-version': [1, 0, 0],
-        'native-assets': toDartConst(),
-      };
-
-  String toNativeAssetsFile() => yamlEncode(toNativeAssetsFileEncoding());
-
-  Future<bool> allExist() async {
-    final allResults = await Future.wait(map((e) => e.exists()));
-    final missing = allResults.contains(false);
-    return !missing;
-  }
-}
-
-Map<X, Y> _combineMaps<X, Y>(Iterable<Map<X, Y>> maps) {
-  final result = <X, Y>{};
-  for (final map in maps) {
-    result.addAll(map);
-  }
-  return result;
 }